Reinit with gitignore
Reinitiated the repository with the proper gitignore. Created a Unity project. created a basic script that printe "Hello World!" to the game console on game start. Imported basic art assets for a character. created a simple 2D platform with collider. Created a basic character with collider and rigidbody2D. added a simple scrip that allow you to scoot the character.