Ground Check
Reed movement script to work with an axis rather than key press. added a ground check to the movement script and adjusted it so the player can only jump when in contact with ground (of course... this includes being beside the ground as well for now...) also imported my rocks and large ground art.